+++ src/main/java/com/jjambbong/note/dto/MemberDto.java
@@ -1,33 +1,13 @@
 package com.jjambbong.note.dto;
 
-import com.jjambbong.note.entity.Member;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
+import lombok.Data;
 
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import java.util.List;
+import java.time.LocalDate;
 
-@Repository
-@RequiredArgsConstructor
+@Data
 public class MemberDto {
-
-    private final EntityManager em;
-
-    // 회원 가입
-    public void save(Member member)  {
-        em.persist(member);
-    }
-
-    // 회원 조회
-    public Member findById(Long id) {
-        return em.find(Member.class, id);
-    }
-
-    // 회원 이메일로 조회
-    public List<Member> findByEmail(String email) {
-        return em.createQuery("select m from Member m where m.email = :email", Member.class)
-                .setParameter("email", email)
-                .getResultList();
-    }
+    private String email;
+    private String password;
+    private String name;
+    private LocalDate createdDate;
 }
