+++ src/main/java/com/jjambbong/note/repository/MemberRepository.java
@@ -1,0 +1,10 @@
+package com.jjambbong.note.repository;
+
+import com.jjambbong.note.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+    Member findByEmail(String email);
+
+}
