+++ src/main/java/com/jjambbong/note/service/MemberService.java
@@ -1,10 +1,6 @@
 package com.jjambbong.note.service;
 
-import com.jjambbong.note.dto.MemberDto;
 import com.jjambbong.note.entity.Member;
-import lombok.AllArgsConstructor;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
