+++ src/main/java/com/jjambbong/note/serviceImpl/MemberServiceImpl.java
@@ -1,30 +1,29 @@
 package com.jjambbong.note.serviceImpl;
 
-import com.jjambbong.note.dto.MemberDto;
+import com.jjambbong.note.repository.MemberRepository;
+import com.jjambbong.note.repository.MemberRepositoryOld;
 import com.jjambbong.note.entity.Member;
 import com.jjambbong.note.service.MemberService;
-import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Transactional
 @Component
 @RequiredArgsConstructor
 public class MemberServiceImpl implements MemberService{
 
-    private final MemberDto memberDto;
+    private final MemberRepository memberRepository;
 
     // 회원 가입
     @Override
     @Transactional
     public Long registerMember(Member member) {
         validateDuplicateMember(member);
-        memberDto.save(member);
+        memberRepository.save(member);
         return member.getMemberId();
     }
 
@@ -32,13 +31,12 @@ public class MemberServiceImpl implements MemberService{
     @Override
     @Transactional(readOnly = true)
     public Member findMember(Long memberId) {
-        return memberDto.findById(memberId);
+        return memberRepository.findById(memberId);
     }
 
     // 중복 회원 검색
     public void validateDuplicateMember(Member member) {
-        System.out.println("중복 회원 검사 시작");
-        List<Member> findMember = memberDto.findByEmail(member.getEmail());
+        Member findMember = memberRepository.findByEmail(member.getEmail());
 
         if (!findMember.isEmpty()) {
             throw new IllegalStateException("이미 존재하는 회원입니다.");
